Integração

+++ app/src/main/java/com/example/library/ConvidadoInsert.kt
@@ -1,7 +1,7 @@
 package com.example.library.data.supabase
 
 data class ConvidadoInsert(
-    val evento_id: Int,
+    val evento_id: Long,
     val nome: String,
     val descricao: String
 )
